+++ src/entities/auth.ts
@@ -1,6 +1,7 @@
 import { FirebaseAuthTypes } from '@react-native-firebase/auth';
 
 export interface IAuthData extends FirebaseAuthTypes.User {
+    admin?: boolean;
     nome?: string;
     success?: boolean;
     token?: string;